Add Collapsable Menu Script
Added a collapsable menu for the automated markdown toc. Converted the
Board Variant file to use the auto toc markdown. Added a js file to do
the collapsing of the menu. Added an entry to the gitignore file so that
local site files in docs don't get checked into git.